# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Playlisted API
  version: 0.2.0
  summary: OpenAPI-first REST API for the Playlisted audio platform.
  description: |
    Contract for the Playlisted backend. The server validates requests and responses
    against this document, and the client SDK is generated from the same source.
servers:
  - url: http://localhost:4000
    description: Local development
security: []
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/UserRole'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/UserStatus'
        - in: query
          name: featured
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: User collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userId}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get user detail
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/playlists/{playlistId}/tags:
    put:
      tags: [Playlists]
      operationId: setPlaylistTags
      summary: Replace all tag associations on a playlist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPlaylistTagsRequest'
      responses:
        '200':
          description: Updated playlist with new tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/tags:
    get:
      tags: [Admin]
      operationId: adminListTags
      summary: List all tags (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: kind
          schema:
            $ref: '#/components/schemas/TagKind'
      responses:
        '200':
          description: Tag list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTagListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Admin]
      operationId: adminCreateTag
      summary: Create a tag (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateTagRequest'
      responses:
        '201':
          description: Created tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/tags/bulk:
    post:
      tags: [Admin]
      operationId: adminBulkCreateTags
      summary: Bulk create tags from a list, skipping duplicates (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBulkCreateTagsRequest'
      responses:
        '201':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBulkCreateTagsResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/tags/{tagId}:
    patch:
      tags: [Admin]
      operationId: adminUpdateTag
      summary: Update a tag (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tagId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateTagRequest'
      responses:
        '200':
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Admin]
      operationId: adminDeleteTag
      summary: Delete a tag (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tagId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/users:
    get:
      tags: [Admin]
      operationId: adminListUsers
      summary: List all users with full details including email (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/UserRole'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/UserStatus'
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/users/{userId}:
    patch:
      tags: [Admin]
      operationId: adminUpdateUser
      summary: Update a user's role, status, or featured flag (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    UserRole:
      type: string
      enum: [LISTENER, CREATOR, EDITOR, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, INVITED]
    PlaylistType:
      type: string
      enum: [PLAYLIST, ALBUM, MIX, PODCAST_CHANNEL, RELEASE]
    RecordingType:
      type: string
      enum: [SONG, PODCAST_EPISODE, MIX_SEGMENT, OTHER]
    Visibility:
      type: string
      enum: [PUBLIC, UNLISTED, PRIVATE]
    PublishStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    TagKind:
      type: string
      enum: [GENRE, MOOD, CATEGORY, SCENE]
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details: {}
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    UserSummary:
      type: object
      required:
        [id, email, username, displayName, role, status, isFeaturedArtist, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        bio:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isFeaturedArtist:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPlaylistSummary:
      type: object
      additionalProperties: false
      required:
        [id, href, ownerId, title, slug, type, visibility, status, featured, isPinnedOnProfile, itemCount, totalDurationSeconds, createdAt, updatedAt]
      properties:
        id:
          type: string
        href:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        featured:
          type: boolean
        isPinnedOnProfile:
          type: boolean
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          additionalProperties: false
          required: [publicPlaylists]
          properties:
            publicPlaylists:
              type: array
              items:
                $ref: '#/components/schemas/UserPlaylistSummary'
    UserListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [email, displayName]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 1
          maxLength: 50
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        passwordHash:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isFeaturedArtist:
          type: boolean
          default: false
    Tag:
      type: object
      additionalProperties: false
      required: [id, name, slug, kind]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        kind:
          $ref: '#/components/schemas/TagKind'
    OwnerSummary:
      type: object
      additionalProperties: false
      required: [id, username, displayName, role]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
          format: uri-reference
        heroImageUrl:
          type: [string, 'null']
          format: uri-reference
        role:
          $ref: '#/components/schemas/UserRole'
...
    PlaylistDetail:
      type: object
      additionalProperties: false
      required:
        [id, href, ownerId, title, slug, type, visibility, status, featured, isPinnedOnProfile, itemCount, totalDurationSeconds, createdAt, updatedAt, owner, recordings]
      properties:
        id:
          type: string
        href:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        featured:
          type: boolean
        isPinnedOnProfile:
          type: boolean
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/RecordingInPlaylist'
...
    RecordingInPlaylist:
      type: object
      additionalProperties: false
      required:
        [id, uploaderId, publishedPlaylistId, title, audioUrl, recordingType, visibility, status, explicit, playCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        uploaderId:
          type: string
        publishedPlaylistId:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        audioUrl:
          type: string
          format: uri-reference
        audioMimeType:
          type: [string, 'null']
        audioBytes:
          type: [integer, 'null']
        durationSeconds:
          type: [integer, 'null']
        artworkUrl:
          type: [string, 'null']
          format: uri-reference
        recordingType:
          $ref: '#/components/schemas/RecordingType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        trackNumber:
          type: [integer, 'null']
        episodeNumber:
          type: [integer, 'null']
        explicit:
          type: boolean
        releaseDate:
          type: [string, 'null']
          format: date-time
        publishedAt:
          type: [string, 'null']
          format: date-time
        playCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        uploader:
          $ref: '#/components/schemas/OwnerSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
    SetPlaylistTagsRequest:
      type: object
      additionalProperties: false
      required: [tagIds]
      properties:
        tagIds:
          type: array
          items:
            type: string
...
    AdminTag:
      type: object
      additionalProperties: false
      required: [id, name, slug, kind, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        kind:
          $ref: '#/components/schemas/TagKind'
        createdAt:
          type: string
          format: date-time
    AdminTagListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminTag'
    AdminCreateTagRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        kind:
          $ref: '#/components/schemas/TagKind'
    AdminUpdateTagRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        kind:
          $ref: '#/components/schemas/TagKind'
    AdminBulkCreateTagItem:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        kind:
          $ref: '#/components/schemas/TagKind'
    AdminBulkCreateTagsRequest:
      type: array
      maxItems: 500
      items:
        $ref: '#/components/schemas/AdminBulkCreateTagItem'
    AdminBulkCreateTagsResponse:
      type: object
      additionalProperties: false
      required: [created, skipped, skippedNames]
      properties:
        created:
          type: integer
        skipped:
          type: integer
        skippedNames:
          type: array
          items:
            type: string
...
    AdminUpdateUserRequest:
      type: object
      additionalProperties: false
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isFeaturedArtist:
          type: boolean